requirements.txt
django
djangorestframework
django-environ
dj-database-url
pre-commit
psycopg2
